\name{SASmdoc-package}
\alias{SASmdoc-package}
\alias{SASmdoc}
\docType{package}
\title{
Eine Sammlung von SAS Makros.
}
\description{
Eine Sammlung von SAS Makros zur Unterstuetzung der Arbeit in der
Vorlesung Advanced Survey Statistis SS 2010.
}
\details{
\tabular{ll}{
Package: \tab SASmdoc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-06-05\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}

}
\author{
Boyko Amarov
}